# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Authentication API",
        "version": "1.0.0",
        "description": "API documentation for client management"
    },
    "paths": {
        "/sign/in": {
            "post": {
                "summary": "Sign in: create a new client",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Sign"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Client created successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/sign/out/{id}": {
            "delete": {
                "summary": "Sign out: delete a client by ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Client deleted successfully"
                    },
                    "404": {
                        "description": "Client not found"
                    }
                }
            }
        },
        "/email/verify": {
            "get": {
                
                "summary": "Verify a client by ID",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Client verified successfully"
                    },
                    "404": {
                        "description": "Client not found"
                    }
                }
            }
        },
        "/log/in": {
            "post": {
                "summary": "Log in: Enable client access to the system",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Log"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Client created successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/log/out/{id}": {
            "put": {
                "summary": "Log out: Disable client access to the system",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Client deleted successfully"
                    },
                    "404": {
                        "description": "Client not found"
                    }
                }
            }
        },
        "/clients": {
            "get": {
                "summary": "Get all clients",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Client"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Client not found"
                    }
                }
            }
        },
        "/clients/{id}": {
            "get": {
                "summary": "Get a client by ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Client"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Client not found"
                    }
                }
            },
            "put": {
                "summary": "Update a client by ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Client"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Client updated successfully"
                    },
                    "404": {
                        "description": "Client not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Client": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "lastname": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    }
                }
            },
            "Sign": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "lastname": {
                        "type": "string"
                    },
                    "username": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                }
            },
            "Log": {
                "type": "object",
                "properties": {
                    "username": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "ip": {
                        "type": "string"
                    }
                }
            },
            "Logs": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "clientId": {
                        "type": "integer"
                    },
                    "ip": {
                        "type": "string"
                    },
                    "type": {
                        "type": "boolean"
                    },
                    "attempt": {
                        "type": "boolean"
                    }
                }
            }
        }
    }
}